BACKEND/requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.0.340
langchain-community==0.0.1
chromadb==0.4.15
sentence-transformers==2.2.2
pandas==2.0.3
openpyxl==3.1.2